move test _helpers in /bin folder
