+++ test/bin/upload.spec.ts
@@ -9,7 +9,7 @@ const execM: jest.Mocked<typeof Exec> = Exec as any;
 import {task} from 'fp-ts/lib/Task';
 import {left2v, leftTask, right2v, taskEither} from 'fp-ts/lib/TaskEither';
 import {Capabilities, capabilities, upload} from '../../src/bin/upload';
-import {result} from '../_helpers';
+import {result} from './_helpers';
 
 afterEach(() => {
   jest.clearAllMocks();
